requirements.txt
boto3
pytesseract
pdf2image
python-dotenv
pytest
dash
pyrhubarb
dash_bootstrap_components
PyYAML